refactor: 아티스트 관련 훅 및 컴포넌트 수정 및 리팩토링
