+++ apps/client/src/pages/my/page/profile/my-profile.tsx
@@ -1,6 +1,6 @@
-import { useSuspenseQuery } from '@tanstack/react-query';
 import { Header, Footer, Spacing } from '@confeti/design-system';
 import { routePath } from '@shared/constants/path';
+
 import { useUserProfile } from '@pages/my/hooks/use-user-info';
 import Box from '@pages/my/components/profile/box';
 import NoArtistSection from '@pages/my/components/artist/no-artist-section';
@@ -8,16 +8,12 @@ import NoConfetiSection from '@pages/my/components/confeti/no-confeti-section';
 import ArtistSection from '@pages/my/components/artist/artist-section';
 import UserInfo from '@pages/my/components/profile/user-info';
 import ConfetiSection from '@pages/my/components/confeti/confeti-section';
-import { USER_QUERY_OPTIONS } from '@shared/apis/user/user-queries';
-import { useMyArtist } from '@pages/my/hooks/use-my-artist';
+import { useMyArtist, useMyConfeti } from '@pages/my/hooks/use-my-favorites';
 
 const MyProfile = () => {
   const profileData = useUserProfile();
   const { data: artistData } = useMyArtist();
-
-  const { data: performanceData } = useSuspenseQuery(
-    USER_QUERY_OPTIONS.FAVORITE_PERFORMANCES(),
-  );
+  const { data: performanceData } = useMyConfeti();
 
   return (
     <>
@@ -33,7 +29,7 @@ const MyProfile = () => {
         showMore={artistData.artists.length > 0}
       >
         {artistData.artists.length > 0 ? (
-          <ArtistSection />
+          <ArtistSection artists={artistData.artists} />
         ) : (
           <NoArtistSection />
         )}
